+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+# Unreleased
+### Changed
+- Convert the Python classifiers to Fedora SPDX expressions (where possible, based on [PEP 639](https://peps.python.org/pep-0639/))
+- When `--fedora-compliant` option is set, a check of the converted license expressions is performed
+to determine whether the license is allowed in Fedora. Data is provided by the [Fedora Legal team](https://gitlab.com/fedora/legal/fedora-license-data/)
+- tomllib library is used to read toml files from Python 3.11 on
+
+
 # [0.6.1] - 2022-08-01
 ### Fixed
 - Include LICENSE-MIT-0 in the sdist and fix license metadata